Implemented deleteStudent when student does not exist

+++ com.examples.school/src/main/java/com/examples/school/controller/SchoolController.java
@@ -30,6 +30,10 @@ public class SchoolController {
 	}
 
 	public void deleteStudent(Student student) {
+		if (studentRepository.findById(student.getId()) == null) {
+			studentView.showError("No existing student with id " + student.getId(), student);
+			return;
+		}
 		studentRepository.delete(student.getId());
 		studentView.studentRemoved(student);
 	}
